refactor(hooks): split useBatches into focused modules behind a stable facade
529-line hook file (2.6x the 200-line hook limit) split by concern:
types, read queries, lifecycle mutations, and content mutations under
src/hooks/batches/. The useBatches facade re-exports everything, so all
ten importers keep their import path.

Removes the legacy useRaiseMaterialRequirement no-op placeholder (zero
callers; useCreateBatchRequirement is the real path).
