+++ src/hooks/useBatches.ts
@@ -1,529 +1,38 @@
-import { useQuery, useMutation, useQueryClient } from "@tanstack/react-query";
-import { supabase } from "@/integrations/supabase/client";
-import { useProfile } from "@/hooks/useProfile";
-import { toast } from "sonner";
-import { useTranslation } from "react-i18next";
-import { QueryKeys } from "@/lib/queryClient";
-import { logger } from '@/lib/logger';
+/**
+ * Public batch hooks API. The implementation is split by concern under
+ * ./batches/ (types, queries, lifecycle mutations, content mutations); this
+ * facade keeps the import path stable for all consumers.
+ *
+ * The legacy no-op `useRaiseMaterialRequirement` placeholder was removed —
+ * use `useCreateBatchRequirement` instead.
+ */
 
-export type BatchType = "laser_nesting" | "tube_batch" | "saw_batch" | "finishing_batch" | "general";
-export type BatchStatus = "draft" | "ready" | "in_progress" | "completed" | "cancelled" | "blocked";
+export type {
+  Batch,
+  BatchOperation,
+  BatchRequirement,
+  BatchStatus,
+  BatchType,
+  CreateBatchInput,
+} from "./batches/types";
 
-export interface Batch {
-  id: string;
-  tenant_id: string;
-  batch_number: string;
-  batch_type: BatchType;
-  cell_id: string;
-  status: BatchStatus;
-  material: string | null;
-  thickness_mm: number | null;
-  estimated_time: number | null;
-  actual_time: number | null;
-  operations_count: number;
-  notes: string | null;
-  nesting_metadata: Record<string, unknown> | null;
-  external_id: string | null;
-  external_source: string | null;
-  created_by: string | null;
-  started_by: string | null;
-  completed_by: string | null;
-  created_at: string;
-  started_at: string | null;
-  completed_at: string | null;
-  updated_at: string | null;
-  nesting_image_url: string | null;
-  layout_image_url: string | null;
-  parent_batch_id: string | null;
-  material_requirement_raised: boolean;
-  material_requirement_metadata: Record<string, unknown> | null;
-  cell?: {
-    id: string;
-    name: string;
-  };
-  created_by_profile?: {
-    full_name: string;
-  };
-}
+export {
+  useBatch,
+  useBatchOperations,
+  useBatchRequirements,
+  useBatches,
+  useSubBatches,
+} from "./batches/queries";
 
-export interface BatchOperation {
-  id: string;
-  batch_id: string;
-  operation_id: string;
-  sequence_in_batch: number | null;
-  quantity_in_batch: number | null;
-  tenant_id: string;
-  created_at: string;
-  operation?: {
-    id: string;
-    operation_name: string;
-    status: string;
-    part?: {
-      id: string;
-      part_number: string;
-      quantity: number;
-      job?: {
-        id: string;
-        job_number: string;
-        customer: string;
-      };
-    };
-  };
-}
+export {
+  useCreateBatch,
+  useDeleteBatch,
+  useUpdateBatch,
+  useUpdateBatchStatus,
+} from "./batches/batchMutations";
 
-export interface BatchRequirement {
-  id: string;
-  batch_id: string;
-  material_name: string;
-  quantity: number;
-  status: 'pending' | 'ordered' | 'received';
-  created_at: string;
-}
-
-export interface CreateBatchInput {
-  batch_number: string;
-  batch_type: BatchType;
-  cell_id: string;
-  material?: string;
-  thickness_mm?: number;
-  estimated_time?: number;
-  notes?: string;
-  nesting_metadata?: Record<string, unknown>;
-  nesting_image_url?: string;
-  layout_image_url?: string;
-  parent_batch_id?: string;
-  operation_ids?: string[];
-}
-
-export function useBatches(filters?: {
-  status?: BatchStatus;
-  batch_type?: BatchType;
-  cell_id?: string;
-}) {
-  const profile = useProfile();
-
-  return useQuery({
-    queryKey: QueryKeys.batches.all(profile?.tenant_id ?? "", filters as Record<string, unknown>),
-    queryFn: async () => {
-      if (!profile?.tenant_id) throw new Error("No tenant_id available");
-      let query = supabase
-        .from("operation_batches")
-        .select(`
-          *,
-          cell:cells(id, name),
-          created_by_profile:profiles!operation_batches_created_by_fkey(full_name)
-        `)
-        .eq("tenant_id", profile.tenant_id)
-        .order("created_at", { ascending: false });
-
-      if (filters?.status) {
-        query = query.eq("status", filters.status);
-      }
-      if (filters?.batch_type) {
-        query = query.eq("batch_type", filters.batch_type);
-      }
-      if (filters?.cell_id) {
-        query = query.eq("cell_id", filters.cell_id);
-      }
-
-      const { data, error } = await query;
-      if (error) throw error;
-      return data as Batch[];
-    },
-    enabled: !!profile?.tenant_id,
-  });
-}
-
-export function useBatch(batchId: string | undefined) {
-  const profile = useProfile();
-
-  return useQuery({
-    queryKey: QueryKeys.batches.detail(batchId ?? ""),
-    queryFn: async () => {
-      if (!batchId || !profile?.tenant_id) return null;
-
-      const { data, error } = await supabase
-        .from("operation_batches")
-        .select(`
-          *,
-          cell:cells(id, name),
-          created_by_profile:profiles!operation_batches_created_by_fkey(full_name),
-          started_by_profile:profiles!operation_batches_started_by_fkey(full_name),
-          completed_by_profile:profiles!operation_batches_completed_by_fkey(full_name)
-        `)
-        .eq("id", batchId)
-        .eq("tenant_id", profile.tenant_id)
-        .single();
-
-      if (error) throw error;
-      return data as unknown as Batch;
-    },
-    enabled: !!batchId && !!profile?.tenant_id,
-  });
-}
-
-export function useSubBatches(batchId: string | undefined) {
-  const profile = useProfile();
-
-  return useQuery({
-    queryKey: QueryKeys.batches.subBatches(batchId ?? "", profile?.tenant_id ?? ""),
-    queryFn: async () => {
-      if (!batchId || !profile?.tenant_id) return [];
-
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any -- operation_batches table not yet in generated types
-      const { data, error } = await (supabase as any)
-        .from("operation_batches")
-        .select("*, cell:cells(id, name)")
-        .eq("parent_batch_id", batchId)
-        .eq("tenant_id", profile.tenant_id)
-        .order("created_at", { ascending: true });
-
-      if (error) throw error;
-      return (data ?? []) as unknown as Batch[];
-    },
-    enabled: !!batchId && !!profile?.tenant_id,
-  });
-}
-
-export function useBatchOperations(batchId: string | undefined) {
-  const profile = useProfile();
-
-  return useQuery({
-    queryKey: QueryKeys.batches.operations(batchId ?? ""),
-    queryFn: async () => {
-      if (!batchId || !profile?.tenant_id) return [];
-
-      const { data, error } = await supabase
-        .from("batch_operations")
-        .select(`
-          *,
-          operation:operations(
-            id,
-            operation_name,
-            status,
-            part:parts(
-              id,
-              part_number,
-              quantity,
-              job:jobs(id, job_number, customer)
-            )
-          )
-        `)
-        .eq("batch_id", batchId)
-        .eq("tenant_id", profile.tenant_id)
-        .order("sequence_in_batch", { ascending: true });
-
-      if (error) throw error;
-      return data as BatchOperation[];
-    },
-    enabled: !!batchId && !!profile?.tenant_id,
-  });
-}
-
-export function useBatchRequirements(batchId: string | undefined) {
-  const profile = useProfile();
-
-  return useQuery({
-    queryKey: QueryKeys.batches.requirements(batchId ?? ""),
-    queryFn: async () => {
-      if (!batchId || !profile?.tenant_id) return [];
-
-      const { data, error } = await supabase
-        .from("batch_requirements")
-        .select("*")
-        .eq("batch_id", batchId)
-        .eq("tenant_id", profile.tenant_id)
-        .order("created_at", { ascending: true });
-
-      if (error) throw error;
-      return (data ?? []) as unknown as BatchRequirement[];
-    },
-    enabled: !!batchId && !!profile?.tenant_id,
-  });
-}
-
-export function useCreateBatch() {
-  const queryClient = useQueryClient();
-  const profile = useProfile();
-  const { t } = useTranslation();
-
-  return useMutation({
-    mutationFn: async (input: CreateBatchInput) => {
-      if (!profile?.tenant_id) throw new Error(t("common.noTenantId"));
-
-      const insertData = {
-          tenant_id: profile.tenant_id,
-          batch_number: input.batch_number,
-          batch_type: input.batch_type,
-          cell_id: input.cell_id,
-          material: input.material,
-          thickness_mm: input.thickness_mm,
-          estimated_time: input.estimated_time,
-          notes: input.notes,
-          nesting_metadata: input.nesting_metadata as unknown,
-          nesting_image_url: input.nesting_image_url,
-          layout_image_url: input.layout_image_url,
-          parent_batch_id: input.parent_batch_id,
-          created_by: profile.id,
-          status: "draft" as const,
-      };
-      const { data: batch, error: batchError } = await supabase
-        .from("operation_batches")
-        .insert(insertData as never)
-        .select()
-        .single();
-
-      if (batchError) throw batchError;
-
-      if (input.operation_ids && input.operation_ids.length > 0) {
-        const batchOperations = input.operation_ids.map((opId, index) => ({
-          tenant_id: profile.tenant_id,
-          batch_id: batch.id,
-          operation_id: opId,
-          sequence_in_batch: index + 1,
-        }));
-
-        const { error: opsError } = await supabase
-          .from("batch_operations")
-          .insert(batchOperations);
-
-        if (opsError) throw opsError;
-      }
-
-      return batch;
-    },
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ["batches"] });
-      toast.success(t("batches.createSuccess"), { description: t("batches.createSuccessDesc") });
-    },
-    onError: (error: Error) => {
-      toast.error(t("common.error"), { description: error.message });
-    },
-  });
-}
-
-export function useUpdateBatch() {
-  const queryClient = useQueryClient();
-  const profile = useProfile();
-  const { t } = useTranslation();
-
-  return useMutation({
-    mutationFn: async ({ id, updates }: { id: string; updates: Partial<Batch> }) => {
-      if (!profile?.tenant_id) throw new Error(t("common.noTenantId"));
-
-      const { data, error } = await supabase
-        .from("operation_batches")
-        .update(updates as never)
-        .eq("id", id)
-        .eq("tenant_id", profile.tenant_id)
-        .select()
-        .single();
-
-      if (error) throw error;
-      return data;
-    },
-    onSuccess: (_, variables) => {
-      queryClient.invalidateQueries({ queryKey: QueryKeys.batches.detail(variables.id) });
-      queryClient.invalidateQueries({ queryKey: ["batches"] });
-      toast.success(t("batches.updateSuccess"));
-    },
-    onError: (error: Error) => {
-      toast.error(t("common.error"), { description: error.message });
-    },
-  });
-}
-
-export function useUpdateBatchStatus() {
-  const queryClient = useQueryClient();
-  const profile = useProfile();
-  const { t } = useTranslation();
-
-  return useMutation({
-    mutationFn: async ({ batchId, status }: { batchId: string; status: BatchStatus }) => {
-      const updates: Record<string, string> = { status };
-
-      if (status === "in_progress" && profile?.id) {
-        updates.started_at = new Date().toISOString();
-        updates.started_by = profile.id;
-      } else if (status === "completed" && profile?.id) {
-        updates.completed_at = new Date().toISOString();
-        updates.completed_by = profile.id;
-      }
-
-      const { data, error } = await supabase
-        .from("operation_batches")
-        .update(updates)
-        .eq("id", batchId)
-        .eq("tenant_id", profile?.tenant_id)
-        .select()
-        .single();
-
-      if (error) throw error;
-      return data;
-    },
-    onMutate: async ({ batchId, status }) => {
-      await queryClient.cancelQueries({ queryKey: QueryKeys.batches.detail(batchId) });
-      const previous = queryClient.getQueryData(QueryKeys.batches.detail(batchId));
-      queryClient.setQueryData(QueryKeys.batches.detail(batchId), (old: Record<string, unknown> | undefined) =>
-        old ? { ...old, status } : old
-      );
-      return { previous, batchId };
-    },
-    onError: (error: Error, _variables, context) => {
-      if (context?.previous) {
-        queryClient.setQueryData(QueryKeys.batches.detail(context.batchId), context.previous);
-      }
-      toast.error(t("common.error"), { description: error.message });
-    },
-    onSettled: (_, _error, variables) => {
-      queryClient.invalidateQueries({ queryKey: ["batches"] });
-      queryClient.invalidateQueries({ queryKey: QueryKeys.batches.detail(variables.batchId) });
-    },
-  });
-}
-
-export function useAddOperationsToBatch() {
-  const queryClient = useQueryClient();
-  const profile = useProfile();
-  const { t } = useTranslation();
-
-  return useMutation({
-    mutationFn: async ({ batchId, operationIds }: { batchId: string; operationIds: string[] }) => {
-      if (!profile?.tenant_id) throw new Error(t("common.noTenantId"));
-
-      const { data: existing } = await supabase
-        .from("batch_operations")
-        .select("sequence_in_batch")
-        .eq("batch_id", batchId)
-        .eq("tenant_id", profile!.tenant_id)
-        .order("sequence_in_batch", { ascending: false })
-        .limit(1);
-
-      const startSequence = (existing?.[0]?.sequence_in_batch || 0) + 1;
-
-      const batchOperations = operationIds.map((opId, index) => ({
-        tenant_id: profile.tenant_id,
-        batch_id: batchId,
-        operation_id: opId,
-        sequence_in_batch: startSequence + index,
-      }));
-
-      const { error } = await supabase
-        .from("batch_operations")
-        .insert(batchOperations);
-
-      if (error) throw error;
-    },
-    onSuccess: (_, variables) => {
-      queryClient.invalidateQueries({ queryKey: QueryKeys.batches.operations(variables.batchId) });
-      queryClient.invalidateQueries({ queryKey: ["batches"] });
-      queryClient.invalidateQueries({ queryKey: QueryKeys.batches.detail(variables.batchId) });
-      toast.success(t("batches.operationsAdded"));
-    },
-    onError: (error: Error) => {
-      toast.error(t("common.error"), { description: error.message });
-    },
-  });
-}
-
-export function useRemoveOperationFromBatch() {
-  const queryClient = useQueryClient();
-  const profile = useProfile();
-  const { t } = useTranslation();
-
-  return useMutation({
-    mutationFn: async ({ batchOperationId }: { batchOperationId: string }) => {
-      const { error } = await supabase
-        .from("batch_operations")
-        .delete()
-        .eq("id", batchOperationId)
-        .eq("tenant_id", profile?.tenant_id);
-
-      if (error) throw error;
-    },
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ["batches"] });
-    },
-    onError: (error: Error) => {
-      toast.error(t("common.error"), { description: error.message });
-    },
-  });
-}
-
-export function useDeleteBatch() {
-  const queryClient = useQueryClient();
-  const profile = useProfile();
-  const { t } = useTranslation();
-
-  return useMutation({
-    mutationFn: async (batchId: string) => {
-      if (!profile?.tenant_id) throw new Error(t("common.noTenantId"));
-
-      await supabase
-        .from("batch_operations")
-        .delete()
-        .eq("batch_id", batchId)
-        .eq("tenant_id", profile.tenant_id);
-
-      const { error } = await supabase
-        .from("operation_batches")
-        .delete()
-        .eq("id", batchId) // RLS checks should handle sub-batches, but user might need to delete them manually or we add cascade
-        .eq("tenant_id", profile.tenant_id);
-
-      if (error) throw error;
-    },
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ["batches"] });
-      toast.success(t("batches.deleteSuccess"));
-    },
-    onError: (error: Error) => {
-      toast.error(t("common.error"), { description: error.message });
-    },
-  });
-}
-
-export function useCreateBatchRequirement() {
-  const queryClient = useQueryClient();
-  const profile = useProfile();
-  const { t } = useTranslation();
-
-  return useMutation({
-    mutationFn: async ({ batchId, materialName, quantity }: { batchId: string; materialName: string; quantity: number }) => {
-      if (!profile?.tenant_id) throw new Error(t("common.noTenantId"));
-
-      const { data, error } = await supabase
-        .from("batch_requirements" as "operation_batches")
-        .insert({
-          tenant_id: profile.tenant_id,
-          batch_id: batchId,
-          material_name: materialName,
-          quantity: quantity,
-          status: 'pending'
-        } as never)
-        .select()
-        .single();
-
-      if (error) throw error;
-      return data;
-    },
-    onSuccess: (_, variables) => {
-      queryClient.invalidateQueries({ queryKey: QueryKeys.batches.requirements(variables.batchId) });
-      toast.success(t("batches.requirementAdded"));
-    },
-    onError: (error: Error) => {
-      toast.error(t("common.error"), { description: error.message });
-    },
-  });
-}
-
-export function useRaiseMaterialRequirement() {
-  // Legacy hook - might want to deprecate or update to use the new table
-  // For now keeping consistent interface but effectively doing nothing if replaced by useCreateBatchRequirement
-  return useMutation({
-    mutationFn: async ({ batchId }: { batchId: string }) => {
-      // Placeholder implementation to satisfy existing calls
-      logger.debug('useBatches', 'Legacy raise requirement called for', batchId);
-    }
-  })
-}
+export {
+  useAddOperationsToBatch,
+  useCreateBatchRequirement,
+  useRemoveOperationFromBatch,
+} from "./batches/batchContentMutations";
